Allow no trusted applications when setting access
The Access field TrustedApplications can now be set to nil or an empty
slice, better matching the SecAccessCreate API.

When nil, access control list is automatically set to the application
creating the item. When an empty slice no applications are trusted for
the item created.

Due to the fact that (Access).Convert() would have returned nil if
len(TrustedApplications) == 0 causing a SIGTRAP, this is a backward-
compatible change to the API. While this doesn't cover all cases (e.g.
providing a list of trusted applications, but while not trusting the
calling application), I believe this will cover many common cases.