secretservice: replace pkg/errors for Go stdlib errors
The pkg/errors dependency was introduced with the secretservice implementation
in 7f2ef9fddce668397ae2791cde4cadd6136c1fbd in March 2019.

go1.13 (September 2019) introduced native support for unwrapping errors, no
longer requiring this dependency to be used, and the pkg/errors module has
been archived (as feature complete).

While pkg/errors does have some advantages (for example, it can provide a
stack trace), this functionality doesn't appear to be used in this module,
and the pkg/errors package is not used in other implementations (for macOS).

This patch removes the dependency, replacing its use for the equivalent
in Go stdlib.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>