secretservice: replace pkg/errors for Go stdlib errors, and fix unhandled error (#111)
* secretservice: fix SecretService.PromptAndWait discarding error

This code was always returning a nil error instead of the error produced
by the org.freedesktop.Secret.Prompt.Prompt call.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

* secretservice: SecretService.openSessionRaw explicitly handle error

pkg/errors' errors.Wrap function implicitly discards nil-errors. While
this is convenient, it also can be err-prone, as this behavior differs
from go stdlib, making it easy to miss conditions where the reader assumes
an error is returned (but in reality no error).

This patch updates the code to explicitly handle non-nil errors to prevent
accidental regressions if this code would be rewritten using go stdlib.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

* secretservice: rename var that shadowed type

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

* secretservice: fix GoDoc comment

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

* secretservice: replace pkg/errors for Go stdlib errors

The pkg/errors dependency was introduced with the secretservice implementation
in 7f2ef9fddce668397ae2791cde4cadd6136c1fbd in March 2019.

go1.13 (September 2019) introduced native support for unwrapping errors, no
longer requiring this dependency to be used, and the pkg/errors module has
been archived (as feature complete).

While pkg/errors does have some advantages (for example, it can provide a
stack trace), this functionality doesn't appear to be used in this module,
and the pkg/errors package is not used in other implementations (for macOS).

This patch removes the dependency, replacing its use for the equivalent
in Go stdlib.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>

---------

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>